Moving code to broader files to allow for usage in other programs (#2996)
* Moving code to broader files to allow for usage in other programs
