+++ bin/s2nd.c
@@ -13,27 +13,14 @@
  * permissions and limitations under the License.
  */
 
-#include <netinet/tcp.h>
-#include <sys/types.h>
-#include <sys/socket.h>
 #include <sys/stat.h>
-#include <sys/ioctl.h>
 #include <sys/mman.h>
-#include <poll.h>
 #include <netdb.h>
-
-#include <stdlib.h>
 #include <signal.h>
 #include <unistd.h>
 #include <fcntl.h>
-#include <string.h>
-#include <signal.h>
-#include <stdio.h>
 #include <getopt.h>
-
 #include <errno.h>
-
-#include <error/s2n_errno.h>
 
 #include <openssl/crypto.h>
 #include <openssl/err.h>
@@ -150,77 +137,7 @@ uint8_t default_ticket_key[32] = {0x07, 0x77, 0x09, 0x36, 0x2c, 0x2e, 0x32, 0xdf
                                   0xb5, 0x0f, 0x9c, 0x31, 0x22, 0xec, 0x84, 0x4a, 0xd7, 0xc2,
                                   0xb3, 0xe5 };
 
-#define MAX_KEY_LEN 32
-#define MAX_VAL_LEN 255
-
-struct session_cache_entry {
-    uint8_t key[MAX_KEY_LEN];
-    uint8_t key_len;
-    uint8_t value[MAX_VAL_LEN];
-    uint8_t value_len;
-};
-
 struct session_cache_entry session_cache[256];
-
-int cache_store_callback(struct s2n_connection *conn, void *ctx, uint64_t ttl, const void *key, uint64_t key_size, const void *value, uint64_t value_size)
-{
-    struct session_cache_entry *cache = ctx;
-
-    POSIX_ENSURE_INCLUSIVE_RANGE(1, key_size, MAX_KEY_LEN);
-    POSIX_ENSURE_INCLUSIVE_RANGE(1, value_size, MAX_VAL_LEN);
-
-    uint8_t idx = ((const uint8_t *)key)[0];
-
-    memcpy(cache[idx].key, key, key_size);
-    memcpy(cache[idx].value, value, value_size);
-
-    cache[idx].key_len = key_size;
-    cache[idx].value_len = value_size;
-
-    return 0;
-}
-
-int cache_retrieve_callback(struct s2n_connection *conn, void *ctx, const void *key, uint64_t key_size, void *value, uint64_t * value_size)
-{
-    struct session_cache_entry *cache = ctx;
-
-    POSIX_ENSURE_INCLUSIVE_RANGE(1, key_size, MAX_KEY_LEN);
-
-    uint8_t idx = ((const uint8_t *)key)[0];
-
-    POSIX_ENSURE(cache[idx].key_len == key_size, S2N_ERR_INVALID_ARGUMENT);
-    POSIX_ENSURE(memcmp(cache[idx].key, key, key_size) == 0, S2N_ERR_INVALID_ARGUMENT);
-    POSIX_ENSURE(*value_size >= cache[idx].value_len, S2N_ERR_INVALID_ARGUMENT);
-
-    *value_size = cache[idx].value_len;
-    memcpy(value, cache[idx].value, cache[idx].value_len);
-
-    for (uint64_t i = 0; i < key_size; i++) {
-        printf("%02x", ((const uint8_t *)key)[i]);
-    }
-    printf("\n");
-
-    return 0;
-}
-
-int cache_delete_callback(struct s2n_connection *conn, void *ctx, const void *key, uint64_t key_size)
-{
-    struct session_cache_entry *cache = ctx;
-
-    POSIX_ENSURE_INCLUSIVE_RANGE(1, key_size, MAX_KEY_LEN);
-
-    uint8_t idx = ((const uint8_t *)key)[0];
-
-    if (cache[idx].key_len != 0) {
-        POSIX_ENSURE(cache[idx].key_len == key_size, S2N_ERR_INVALID_ARGUMENT);
-        POSIX_ENSURE(memcmp(cache[idx].key, key, key_size) == 0, S2N_ERR_INVALID_ARGUMENT);
-    }
-
-    cache[idx].key_len = 0;
-    cache[idx].value_len = 0;
-
-    return 0;
-}
 
 /*
  * Since this is a server, and the mechanism for hostname verification is not defined for this use-case,
